refactor: restructure DIRACCommon to preserve DIRAC directory hierarchy
Reorganize DIRACCommon modules to match the original DIRAC structure:
- Core/Utilities/ClassAd/ClassAdLight.py (utilities belong in Core)
- WorkloadManagementSystem/DB/JobDBUtils.py (preserves original location)
- WorkloadManagementSystem/Utilities/ParametricJob.py (preserves original location)

Update all imports across DIRAC, diracx, and DIRACCommon to use the new
structured paths. Reorganize tests to match the new directory structure.

This maintains the logical organization while enabling proper code migration.
