+++ dirac-common/src/DIRACCommon/Core/Utilities/ClassAd/__init__.py
@@ -1,0 +1,1 @@
+"""ClassAd utilities for DIRACCommon"""
