+++ dirac-common/src/DIRACCommon/Core/Utilities/DErrno.py
