+++ dirac-common/src/DIRACCommon/Core/__init__.py
@@ -1,0 +1,1 @@
+"""DIRACCommon Core utilities"""
