--- dirac-common/src/DIRACCommon/Utils/ClassAd/__init__.py
@@ -1,1 +1,0 @@
-"""ClassAd utilities for DIRACCommon"""
