+++ dirac-common/src/DIRACCommon/WorkloadManagementSystem/DB/__init__.py
@@ -1,0 +1,1 @@
+"""DIRACCommon WorkloadManagementSystem DB utilities"""
