+++ dirac-common/src/DIRACCommon/WorkloadManagementSystem/Utilities/__init__.py
@@ -1,0 +1,1 @@
+"""DIRACCommon WorkloadManagementSystem utilities"""
