+++ dirac-common/tests/Core/Utilities/ClassAd/test_ClassAdLight.py
@@ -4,17 +4,17 @@
 
 import pytest
 
-from DIRACCommon.Utils.ClassAd.ClassAdLight import ClassAd
+from DIRACCommon.Core.Utilities.ClassAd.ClassAdLight import ClassAd
 
 
 def test_classad_basic():
     """Test basic ClassAd functionality"""
     # Arrange
-    jdl = "[Executable = \"my_executable\"; JobName = \"Test\"]"
-    
+    jdl = '[Executable = "my_executable"; JobName = "Test"]'
+
     # Act
     classad = ClassAd(jdl)
-    
+
     # Assert
     assert classad.isOK()
     assert classad.getAttributeString("Executable") == "my_executable"
@@ -26,12 +26,12 @@ def test_classad_insert_attributes():
     # Arrange
     jdl = "[]"
     classad = ClassAd(jdl)
-    
+
     # Act
     classad.insertAttributeString("Executable", "test_exe")
     classad.insertAttributeInt("Priority", 5)
     classad.insertAttributeBool("Test", True)
-    
+
     # Assert
     assert classad.getAttributeString("Executable") == "test_exe"
     assert classad.getAttributeInt("Priority") == 5
@@ -41,12 +41,12 @@ def test_classad_insert_attributes():
 def test_classad_list_attributes():
     """Test list attributes in ClassAd"""
     # Arrange
-    jdl = "[InputData = {\"file1.txt\", \"file2.txt\"}]"
+    jdl = '[InputData = {"file1.txt", "file2.txt"}]'
     classad = ClassAd(jdl)
-    
+
     # Act
     file_list = classad.getListFromExpression("InputData")
-    
+
     # Assert
     assert len(file_list) == 2
     assert "file1.txt" in file_list
@@ -56,12 +56,12 @@ def test_classad_list_attributes():
 def test_classad_as_jdl():
     """Test converting ClassAd back to JDL"""
     # Arrange
-    original_jdl = "[Executable = \"my_executable\"; JobName = \"Test\"]"
+    original_jdl = '[Executable = "my_executable"; JobName = "Test"]'
     classad = ClassAd(original_jdl)
-    
+
     # Act
     regenerated_jdl = classad.asJDL()
-    
+
     # Assert
     assert "Executable" in regenerated_jdl
     assert "my_executable" in regenerated_jdl
