--- dirac-common/tests/Utils/__init__.py
@@ -1,1 +1,0 @@
-"""DIRACCommon Utils tests"""
