+++ src/DIRAC/Core/Utilities/ClassAd/ClassAdLight.py
@@ -3,7 +3,7 @@
 """
 
 # Import from DIRACCommon for backward compatibility
-from DIRACCommon.Utils.ClassAd.ClassAdLight import ClassAd
+from DIRACCommon.Core.Utilities.ClassAd.ClassAdLight import ClassAd
 
 # Re-export for backward compatibility
 __all__ = ["ClassAd"]
